requirements.txt
pyparted==3.13.0
